requirements.txt
# Super-Codex-AI Requirements
# Core FastAPI dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.0
starlette==0.38.6
pydantic==2.9.0
pydantic-settings==2.12.0
orjson==3.11.4
jinja2==3.1.4
httpx==0.28.1

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg==3.2.3

# Additional utilities
python-multipart==0.0.6
email-validator==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Monitoring
prometheus-fastapi-instrumentator==7.0.0
sentry-sdk[fastapi]==1.40.0

# System monitoring and service management
psutil>=5.9.6

# HTTP clients and web utilities
requests>=2.31.0
aiohttp>=3.9.1

# Configuration and environment
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Authentication and security
pyjwt>=2.8.0
cryptography>=41.0.0

# File handling
pillow>=10.1.0
python-magic>=0.4.27

# Utility libraries
click>=8.1.7
rich>=13.7.0
colorama>=0.4.6

# WooCommerce integration
woocommerce>=3.0.0

# Axiom Flame dependencies
flask>=3.0.0
# JSON schema validation
jsonschema>=4.19.0

# Additional dependencies for missing modules
schedule>=1.2.0
waitress>=2.1.2

# Enhanced JSON and data handling
bleach>=6.1.0
markupsafe>=2.1.3
unidecode>=1.3.7

# Development and testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code quality
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0

